LayoutTests/imported/w3c/web-platform-tests/css/css-values/tree-counting/sibling-index-keyframe-font-variation-settings-dynamic-expected.txt


PASS Initially, the sibling-index() is 3 for #target
FAIL Removing a preceding sibling of #target reduces the sibling-index() assert_equals: expected "\"wght\" 2" but got "\"wght\" 3"
